[Fix] way of storing cart

+++ components/cartproduct.js
@@ -12,29 +12,48 @@ initFirebase()
 
 const CartProduct = props => {
     const [product, setProduct] = useState([])
-    const { cartState, setCartState } = useContext(CartContext)
+    const { cartContext, setCartContext } = useContext(CartContext)
     const { totalPrice, setTotalPrice } = useContext(TotalPriceContext)
 
     useEffect(() => {
         firebase.firestore().collection("products").doc(props.dbId).get().then(doc => {
             setProduct(doc.data())
-            console.log(doc.data())
             setTotalPrice(doc.data().price * props.quantity)
         })
     }, [])
 
     const removeItem = () => {
         if (props.authId) {
-            firebase.firestore().collection(`users/${props.authId}/cart`).doc(props.dbId).delete().then(() => {
-                setTotalPrice(-(product.price * props.quantity))
-                setCartState(true)
+            firebase.firestore().collection(`users/${props.authId}/cart`).doc(props.dbId).get().then(doc => {
+                if (doc.exists) {
+                    if (doc.data().quantity > 1) {
+                        firebase.firestore().collection(`users/${props.authId}/cart`).doc(props.dbId).update({
+                            quantity: doc.data().quantity - 1
+                        }).then(() => {
+                            setTotalPrice(-(product.price * props.quantity))
+                            setCartContext(true)
+                        })
+                    } else {
+                        firebase.firestore().collection(`users/${props.authId}/cart`).doc(props.dbId).delete().then(() => {
+                            setTotalPrice(-(product.price * props.quantity))
+                            setCartContext(true)
+                        })
+                    }
+                }
             })
+
         } else {
             let cart = JSON.parse(localStorage.getItem("cart"))
-            delete cart[props.dbId]
+            if (cart[props.dbId]) {
+                if (cart[props.dbId].quantity > 1) {
+                    cart[props.dbId].quantity--
+                } else {
+                    delete cart[props.dbId]
+                }
+            }
             localStorage.setItem("cart", JSON.stringify(cart))
             setTotalPrice(-(product.price * props.quantity))
-            setCartState(true)
+            setCartContext(true)
         }
     }
